+++ src/interface/meshblock_data.hpp
@@ -13,9 +13,11 @@
 #ifndef INTERFACE_MESHBLOCK_DATA_HPP_
 #define INTERFACE_MESHBLOCK_DATA_HPP_
 
+#include <algorithm>
 #include <limits>
 #include <map>
 #include <memory>
+#include <set>
 #include <string>
 #include <utility>
 #include <vector>
@@ -109,8 +111,49 @@ class MeshBlockData {
   /// Create copy of MeshBlockData, possibly with a subset of named fields,
   /// and possibly shallow.  Note when shallow=false, new storage is allocated
   /// for non-OneCopy vars, but the data from src is not actually deep copied
-  void Initialize(const MeshBlockData<T> *src, const std::vector<std::string> &names,
-                  const bool shallow);
+  template <typename ID_t>
+  void Initialize(const MeshBlockData<T> *src, const std::vector<ID_t> &vars,
+                  const bool shallow_copy) {
+    PARTHENON_DEBUG_REQUIRE(src != nullptr, "Source data must be non-null.");
+    SetBlockPointer(src);
+    resolved_packages_ = src->resolved_packages_;
+    is_shallow_ = shallow_copy;
+    
+    // clear all variables, maps, and pack caches
+    varVector_.clear();
+    varMap_.clear();
+    varUidMap_.clear();
+    flagsToVars_.clear();
+    varPackMap_.clear();
+    coarseVarPackMap_.clear();
+    varFluxPackMap_.clear();
+
+    auto add_var = [=](auto var) {
+      if (shallow_copy || var->IsSet(Metadata::OneCopy)) {
+        Add(var);
+      } else {
+        Add(var->AllocateCopy(pmy_block));
+      }
+    };
+
+    // special case when the list of vars is empty, copy everything
+    if (vars.empty()) {
+      for (auto v : src->GetVariableVector()) {
+        add_var(v);
+      }
+    } else {
+      for (const auto &v : vars) {
+        auto &var = src->GetVarPtr(v);
+        add_var(src->GetVarPtr(var));
+        // Add the associated flux as well if not explicitly
+        // asked for
+        if (var->metadata().GetFluxName() != "" &&
+          std::find(names.begin(), names.end(), var->metadata().GetFluxName()) ==
+              names.end())
+        add_var(var_map.at(var->metadata().GetFluxName()));
+      }
+    }
+  }
 
   //
   // Queries related to Variable objects
@@ -124,13 +167,13 @@ class MeshBlockData {
   const MapToVars<T> &GetVariableMap() const noexcept { return varMap_; }
 
   std::shared_ptr<Variable<T>> GetVarPtr(const std::string &label) const {
-    PARTHENON_REQUIRE(varMap_.count(label), "Asking for variable " + label +
-                                                " that is not in this MeshBlockData.");
+    PARTHENON_REQUIRE_THROWS(varMap_.count(label),
+                             "Couldn't find variable '" + label + "'");
     return varMap_.at(label);
   }
   std::shared_ptr<Variable<T>> GetVarPtr(const Uid_t &uid) const {
-    PARTHENON_REQUIRE(varUidMap_.count(uid),
-                      "Variable ID " + std::to_string(uid) + "not found!");
+    PARTHENON_REQUIRE_THROWS(varUidMap_.count(uid),
+                             "Variable ID " + std::to_string(uid) + " not found!");
     return varUidMap_.at(uid);
   }
 
@@ -390,15 +433,18 @@ class MeshBlockData {
     return (my_keys == cmp_keys);
   }
 
-  bool Contains(const std::string &name) const noexcept {
-    if (varMap_.find(name) != varMap_.end()) return true;
-    return false;
+  bool Contains(const std::string &name) const noexcept { return varMap_.count(name); }
+  bool Contains(const Uid_t &uid) const noexcept { return varUidMap_.count(uid); }
+  template <typename ID_t>
+  bool Contains(const std::vector<ID_t> &vars) const noexcept {
+    return std::all_of(vars.begin(), vars.end(),
+                       [this](const auto &v) { return this->Contains(v); });
   }
-  bool Contains(const std::vector<std::string> &names) const noexcept {
-    for (const auto &name : names) {
-      if (!Contains(name)) return false;
-    }
-    return true;
+  template <typename ID_t>
+  bool ContainsExactly(const std::vector<ID_t> &vars) const noexcept {
+    // JMM: Assumes vars contains no duplicates. But that would have
+    // been caught elsewhere because `MeshBlockData::Add` would have failed.
+    return Contains(vars) && (vars.size() == varVector_.size());
   }
 
   void SetAllVariablesToInitialized() {
@@ -421,8 +467,9 @@ class MeshBlockData {
                 int sparse_id = InvalidSparseID);
 
   void Add(std::shared_ptr<Variable<T>> var) noexcept {
-    PARTHENON_REQUIRE(varMap_.count(var->label()) == 0,
-                      "Trying to add the same variable twice.");
+    if (varUidMap_.count(var->GetUniqueID())) {
+      PARTHENON_THROW("Tried to add variable " + var->label() + " twice!");
+    }
     varVector_.push_back(var);
     varMap_[var->label()] = var;
     varUidMap_[var->GetUniqueID()] = var;